#TK-2242 "Bug do Milenio" Alteração de Int32 para Int64 para o idLote
git-svn-id: svn://svn.code.sf.net/p/acbr/code/trunk2@24135 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Fontes/ACBrDFe/ACBrBPe/ACBrBPe.pas
@@ -95,7 +95,7 @@ type
 
     function Cancelamento(const AJustificativa: String; ALote: Integer = 0): Boolean;
     function Consultar(const AChave: String = ''; AExtrairEventos: Boolean = False): Boolean;
-    function EnviarEvento(idLote: Integer): Boolean;
+    function EnviarEvento(idLote: Int64): Boolean;
 
     procedure LerServicoDeParams(LayOutServico: TLayOutBPe; var Versao: Double;
       var URL: String); reintroduce; overload;
@@ -518,7 +518,7 @@ begin
   end;
 end;
 
-function TACBrBPe.EnviarEvento(idLote: Integer): Boolean;
+function TACBrBPe.EnviarEvento(idLote: Int64): Boolean;
 var
   i, j: Integer;
   chBPe: String;
