+++ Fontes/ACBrDFe/ACBrBPe/ACBrBPeWebServices.pas
@@ -215,7 +215,7 @@ type
 
   TBPeEnvEvento = class(TBPeWebService)
   private
-    FidLote: Integer;
+    FidLote: Int64;
     FEvento: TEventoBPe;
     FcStat: Integer;
     FxMotivo: String;
@@ -238,7 +238,7 @@ type
     destructor Destroy; override;
     procedure Clear; override;
 
-    property idLote: Integer read FidLote write FidLote;
+    property idLote: Int64 read FidLote write FidLote;
     property cStat: Integer read FcStat;
     property xMotivo: String read FxMotivo;
     property TpAmb: TpcnTipoAmbiente read FTpAmb;
