+++ Fontes/ACBrDFe/ACBrMDFe/ACBrMDFeWebServices.pas
@@ -319,7 +319,7 @@ type
 
   TMDFeEnvEvento = class(TMDFeWebService)
   private
-    FidLote: Integer;
+    FidLote: Int64;
     FEvento: TEventoMDFe;
     FcStat: Integer;
     FxMotivo: String;
@@ -343,7 +343,7 @@ type
     destructor Destroy; override;
     procedure Clear; override;
 
-    property idLote: Integer read FidLote write FidLote;
+    property idLote: Int64 read FidLote write FidLote;
     property cStat: Integer read FcStat;
     property xMotivo: String read FxMotivo;
     property TpAmb: TpcnTipoAmbiente read FTpAmb;
