+++ Fontes/ACBrDFe/ACBrMDFe/PCNMDFe/pmdfeEnvEventoMDFe.pas
@@ -78,7 +78,7 @@ type
   TEventoMDFe = class(TObject)
   private
     FGerador: TGerador;
-    FidLote: Integer;
+    FidLote: Int64;
     FEvento: TInfEventoCollection;
     FVersao: String;
     
@@ -94,7 +94,7 @@ type
     function LerFromIni(const AIniString: String): Boolean;
 
     property Gerador: TGerador             read FGerador write FGerador;
-    property idLote: Integer               read FidLote  write FidLote;
+    property idLote: Int64                 read FidLote  write FidLote;
     property Evento: TInfEventoCollection  read FEvento  write SetEvento;
     property Versao: String                read FVersao  write FVersao;
   end;
@@ -485,7 +485,7 @@ begin
   try
     LerIniArquivoOuString(AIniString, INIRec);
 
-    idLote := INIRec.ReadInteger('EVENTO', 'idLote', 0);
+    idLote := INIRec.ReadInt64('EVENTO', 'idLote', 0);
 
     I := 1;
     while true do
