+++ Fontes/ACBrDFe/ACBrNF3e/PCNNF3e/pcnEnvEventoNF3e.pas
@@ -77,7 +77,7 @@ type
   TEventoNF3e = class(TObject)
   private
     FGerador: TGerador;
-    FidLote: Integer;
+    FidLote: Int64;
     FEvento: TInfEventoCollection;
     FVersao: String;
 
@@ -92,7 +92,7 @@ type
     function LerFromIni(const AIniString: String; CCe: Boolean = True): Boolean;
 
     property Gerador: TGerador            read FGerador write FGerador;
-    property idLote: Integer              read FidLote  write FidLote;
+    property idLote: Int64                read FidLote  write FidLote;
     property Evento: TInfEventoCollection read FEvento  write SetEvento;
     property Versao: String               read FVersao  write FVersao;
   end;
@@ -361,7 +361,7 @@ begin
   INIRec := TMemIniFile.Create('');
   try
     LerIniArquivoOuString(AIniString, INIRec);
-    idLote := INIRec.ReadInteger( 'EVENTO', 'idLote', 0);
+    idLote := INIRec.ReadInt64('EVENTO', 'idLote', 0);
 
     I := 1 ;
     while true do
